apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pymysql==1.1.1
alembic==1.14.0
pydantic-settings==2.6.1
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-docx==1.1.2
weasyprint==63.0
reportlab==4.2.5
openai==1.57.0
anthropic==0.42.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
